requirements.txt
python-dotenv
influxdb-client
pandas
numpy
ccxt
ta
requests
